+++ Dockerfile
@@ -3,7 +3,7 @@ RUN mkdir /data
 # tz data will prompt for interactive Q&A which we dont want
 ARG DEBIAN_FRONTEND=noninteractive
 # INSTALL PYTHON AND DEPENDENCIES
-RUN apt-get -y update && apt-get install -y build-essential python3.6 python3-pip python3-dev cython3 checkinstall cmake default-jdk git libreadline-gplv2-dev libncursesw5-dev libssl-dev libsqlite3-dev tk-dev libgdbm-dev libc6-dev liblcms2-dev libtiff-dev libpng-dev libvips libvips-dev libz-dev libbz2-dev libjpeg-dev openslide-tools unzip python3-openslide wget xutils-dev zlib1g-dev libopenjp2-7-dev  python3-gdcm numpy
+RUN apt-get -y update && apt-get install -y build-essential python3.6 python3-pip python3-dev cython3 checkinstall cmake default-jdk git libreadline-gplv2-dev libncursesw5-dev libssl-dev libsqlite3-dev tk-dev libgdbm-dev libc6-dev liblcms2-dev libtiff-dev libpng-dev libvips libvips-dev libz-dev libbz2-dev libjpeg-dev openslide-tools unzip python3-openslide wget xutils-dev zlib1g-dev libopenjp2-7-dev  python3-gdcm
 RUN pip3 install git+https://github.com/Who8MyLunch/CharPyLS
 
 RUN git clone https://github.com/Steven-N-Hart/dicom_wsi.git
@@ -11,7 +11,7 @@ RUN cd dicom_wsi && pip3 install -r requirements.txt && cd /
 ENV PYTHONPATH=$PYTHONPATH:/dicom_wsi/dicom_wsi/mods/
 #
 ## Add other tools
-##Google
+## Google
 #RUN wget https://github.com/GoogleCloudPlatform/wsi-to-dicom-converter/releases/download/v1.0.3/wsi2dcm_1.0.3.deb
 #RUN apt install ./wsi2dcm_1.0.3.deb
 #
