+++ dicom_wsi/__init__.py
@@ -2,4 +2,5 @@
 """Top-level package for dicom-wsi."""
 __author__ = """Steven N. Hart"""
 __email__ = 'steven.n.hart@gmail.com'
-__version__ = '0.1.0'
+__version__ = '1.0.0'
+
