+++ dicom_wsi/character_validations.py
