+++ dicom_wsi/cli.py
@@ -6,7 +6,7 @@ import logging
 import os
 import sys
 
-from dicom_wsi.mods import parse_wsi
+from . import parse_wsi
 from yaml import load, BaseLoader
 
 import dicom_wsi
