+++ dicom_wsi/extract_annotations.py
