+++ dicom_wsi/extract_image_patches.py
