+++ dicom_wsi/image_filter.py
