+++ dicom_wsi/run.py
@@ -1,13 +1,13 @@
 import logging
 from timeit import default_timer as timer
 import os
-from dicom_wsi.mods.base_attributes import build_base
-from dicom_wsi.mods.parse_wsi import get_wsi
-from dicom_wsi.mods.pixel_data_conversion import resize_wsi_image
-from dicom_wsi.mods.pixel_to_slide_conversions import add_per_frame_functional_groups_sequence
-from dicom_wsi.mods.sequence_attributes import build_sequences
-from dicom_wsi.mods.shared_functional_groups import build_functional_groups
-from dicom_wsi.mods.add_annotations import add_annotations
+from .base_attributes import build_base
+from .parse_wsi import get_wsi
+from .pixel_data_conversion import resize_wsi_image
+from .pixel_to_slide_conversions import add_per_frame_functional_groups_sequence
+from .sequence_attributes import build_sequences
+from .shared_functional_groups import build_functional_groups
+from .add_annotations import add_annotations
 
 logger = logging.getLogger(__name__)
