+++ dicom_wsi/sequence_attributes.py
