+++ dicom_wsi/shared_functional_groups.py
