+++ setup.py
@@ -46,7 +46,7 @@ setup(
     include_package_data=True,
     keywords='dicom_wsi',
     name='dicom_wsi',
-    packages=find_packages(include=['dicom_wsi', 'dicom_wsi.*']),
+    packages=find_packages(),
     setup_requires=setup_requirements,
     test_suite='tests',
     tests_require=test_requirements,
