+++ tests/test_base_attributes.py
@@ -7,8 +7,8 @@ import os
 
 from yaml import load, BaseLoader
 
-from ..dicom_wsi.mods.base_attributes import build_base
-from ..dicom_wsi.mods.parse_wsi import *
+from ..dicom_wsi.base_attributes import build_base
+from ..dicom_wsi.parse_wsi import get_wsi
 
 
 def test_get_wsi():
