+++ tests/test_character_validations.py
@@ -2,13 +2,9 @@
 # -*- coding: utf-8 -*-
 
 """Tests for `dicom_wsi` package."""
-import os
-import sys
-
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 import pytest
+from ..dicom_wsi import character_validations as cv
 
-from ..dicom_wsi.mods import character_validations as cv
 
 def test_cs():
     cv.cs_validator('patient', 'Roberto')
